Refactor sub interpreter handling, pushing into Main Go routine.
This cleans up some stuff, while (temporarily) complicating others.

It's important to note that the main interpreter MUST release
itself when it can...so even using the "global" interpreter we
must do a dance because there *might* be other non-global
interpreters. This could be optimized later to check a boolean or
something.