Solve Python deadlocks by only executing code on global interpreter (#1)
When not using sub-interpreters (e.g. mode isolated or isolated_legacy), certain Python modules that do lots of funky stuff behind the scenes using native libraries can cause frustrating-to-debug deadlocks in the Python interpreter integration.

This redesign moves all execution of global interpreter Python code to the "main" Go routine that manages the interpreter. This seems to fix the deadlocks and allows for dropping the modules: [] config property, too.

This does make it look like we need a future rewrite to clean up and merge more of the overlap between the single and multi-interpreter code.